refactor: extract system monitor layout helpers from main
Made-with: Cursor
